Here are the issues I have found so far:

#### Problems with type parameters:

Trying to use a function that has a type parameter in its return value (there are some debug statements in here).:

```jl
julia>  @interpret Base.unsafe_convert(Ptr{Float64}, rand(5))
(RetType, ArgType) = (Ptr{T}, svec(Any))
def = :(function ccall_jl_array_ptr(arg1, ::Type{T}) where T
      return ccall(:jl_array_ptr, Ptr{T}, (Any,), arg1)
  end)
Calling compiled function JuliaInterpreter.CompiledCalls.ccall_jl_array_ptr with args: Any[[0.0503368, 0.692781, 0.877826, 0.0605047, 0.586223], Float64]
ERROR: ccall return type Ptr should have an element type (not Ptr{<:T})
```

This is what I meant with the comment at https://github.com/JuliaDebug/JuliaInterpreter.jl/issues/206#issuecomment-476583694.

Copy pasting and this works:

```jl
julia> def = :(function ccall_jl_array_ptr(arg1, ::Type{T}) where T
             return ccall(:jl_array_ptr, Ptr{T}, (Any,), arg1)
         end);

julia> eval(def)
ccall_jl_array_ptr (generic function with 1 method)

julia> ccall_jl_array_ptr([1,2], Int64)
Ptr{Int64} @0x0000000008391250
```

#### Failure with toplevel things:

```jl
julia> ex = quote   # in lowered code, cf is a Symbol
           cf = @eval @cfunction(fcfun, Int, (Int, Int))
           ccall(cf, Int, (Int, Int), 1, 2)
       end;

julia> frame = JuliaInterpreter.prepare_thunk(Main, ex)
ERROR: UndefVarError: cf not defined
Stacktrace:
 [1] macro expansion at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\interpret.jl:64 [inlined]
 [2] #build_compiled_call!#47(::Symbol, ::Function, ::Expr, ::Symbol, ::Symbol, ::Array{Any,1}, ::Core.CodeInfo, ::Int64, ::Int64, ::Array{Symbol,1}) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\optimize.jl:209
 [3] build_compiled_call!(::Expr, ::Symbol, ::Symbol, ::Array{Any,1}, ::Core.CodeInfo, ::Int64, ::Int64, ::Array{Symbol,1}) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\optimize.jl:190
 [4] optimize!(::Core.CodeInfo, ::Module, ::Array{Symbol,1}) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\optimize.jl:178
 [5] #FrameCode#3(::Bool, ::Bool, ::Type, ::Module, ::Core.CodeInfo) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\types.jl:71
 [6] JuliaInterpreter.FrameCode(::Module, ::Core.CodeInfo) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\types.jl:69
 [7] prepare_thunk(::Module, ::Expr, ::Bool) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\construct.jl:322
 [8] prepare_thunk(::Module, ::Expr, ::Bool) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\construct.jl:335
 [9] prepare_thunk(::Module, ::Expr) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\construct.jl:321
 [10] top-level scope at none:0
```

#### Some other failure:


```jl
julia> ex = quote
           length(code_typed(fcfun, (Int, Int)))
       end;

julia> fcfun(x::Int, y::Int) = 1;

julia> frame = JuliaInterpreter.prepare_thunk(Main, ex);

julia>  JuliaInterpreter.finish_and_return!(frame, true);
(RetType, ArgType) = (Bool, svec())
def = :(function $(Expr(:where, :(ccall_jl_is_in_pure_context())))
      #= C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\optimize.jl:231 =#
      return ccall(:jl_is_in_pure_context, Bool, ())
  end)
Calling compiled function JuliaInterpreter.CompiledCalls.ccall_jl_is_in_pure_context with args: Any[]
(RetType, ArgType) = (Ptr{Nothing}, svec(Any))
def = :(function $(Expr(:where, :(ccall_jl_value_ptr(arg1))))
      #= C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\optimize.jl:231 =#
      return ccall(:jl_value_ptr, Ptr{Nothing}, (Any,), arg1)
  end)
Calling compiled function JuliaInterpreter.CompiledCalls.ccall_jl_value_ptr with args: Any[svec(Int64, Int64)]
Calling compiled function JuliaInterpreter.CompiledCalls.ccall_jl_value_ptr with args: Any[svec(Int64, Int64)]
Calling compiled function JuliaInterpreter.CompiledCalls.ccall_jl_value_ptr with args: Any[svec(Int64, Int64)]
Calling compiled function JuliaInterpreter.CompiledCalls.ccall_jl_value_ptr with args: Any[Ptr{Nothing} @0x0000000005ad07f0]
ERROR: argument tuple type must contain only types
Stacktrace:
 [1] maybe_evaluate_builtin(::Frame, ::Expr, ::Bool) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\builtins-julia1.1.jl:111
 [2] #evaluate_call_recurse!#37(::Bool, ::Function, ::Any, ::Frame, ::Expr) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\interpret.jl:208
 [3] evaluate_call_recurse! at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\interpret.jl:205 [inlined]
 [4] eval_rhs(::Any, ::Frame, ::Expr) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\interpret.jl:371
 [5] step_expr!(::Any, ::Frame, ::Any, ::Bool) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\interpret.jl:504
 [6] step_expr! at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\interpret.jl:543 [inlined]
 [7] finish!(::Any, ::Frame, ::Bool) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\commands.jl:14
 [8] finish_and_return! at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\commands.jl:29 [inlined]
 [9] #evaluate_call_recurse!#37(::Bool, ::Function, ::Any, ::Frame, ::Expr) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\interpret.jl:242
```