@KristofferC 2019-03-26
The positive thing

```
julia> const s = "foo";

julia>  @btime @interpret pointer_from_objref(s)
  4.886 μs (44 allocations: 2.38 KiB)

```

vs master

```
julia> @btime @interpret pointer_from_objref(s)
  1.170 ms (503 allocations: 30.91 KiB)
```