@KristofferC 2019-03-26
Both of these calls exist in Base:

```
ccall(:jl_value_ptr, Any, (Ptr{Cvoid},), x)
ccall(:jl_value_ptr, Ptr{Cvoid}, (Any,), x)
```

so we cannot just name the wrapper based on the function itself.