@KristofferC 2019-03-26
Fixed some stuff and upgraded to a segfault while running the tests.