@timholy 2019-03-27
Looks like you've taken this quite far, and many of the former problems are gone.

I inserted some debugging statements. Ran out of time to get any further, but for the out-of-memory-error I saw this during the initial construction:

```
stmt = :($(Expr(:foreigncall, :(:jl_array_grow_end), Nothing, svec(Any, UInt64), :(:ccall), 2, :(_2), :(%2), :(%1))))
methname = :ccall_70e68ad9_99df_4f44_8dfc_af16c88206f5
def = :(function $(Expr(:where, :(ccall_70e68ad9_99df_4f44_8dfc_af16c88206f5(arg1, arg2))))
      #= /home/tim/.julia/dev/JuliaInterpreter/src/optimize.jl:241 =#
      #= /home/tim/.julia/dev/JuliaInterpreter/src/optimize.jl:241 =# @show ccall_70e68ad9_99df_4f44_8dfc_af16c88206f5 arg1 arg2
      #= /home/tim/.julia/dev/JuliaInterpreter/src/optimize.jl:242 =#
      return ccall(:jl_array_grow_end, Nothing, (Any, UInt64), arg1, arg2)
  end)
stmt pre: $(Expr(:foreigncall, :(:jl_array_grow_end), Nothing, svec(Any, UInt64), :(:ccall), 2, :(_2), :(%2), :(%1)))
stmt post: ($(QuoteNode(JuliaInterpreter.CompiledCalls.ccall_70e68ad9_99df_4f44_8dfc_af16c88206f5)))(_2, %2)
ccall_70e68ad9_99df_4f44_8dfc_af16c88206f5 = JuliaInterpreter.CompiledCalls.ccall_70e68ad9_99df_4f44_8dfc_af16c88206f5
arg1 = UInt8[]
arg2 = 0x0000000000000001
```

but for the one that causes the error

```
ccall_70e68ad9_99df_4f44_8dfc_af16c88206f5 = JuliaInterpreter.CompiledCalls.ccall_70e68ad9_99df_4f44_8dfc_af16c88206f5
arg1 = UInt8[0x66, 0x65, 0x76, 0x61, 0x6c, 0x5f, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]
arg2 = 0x00007fffafa3f63f
```

That's rather a lot of elements! Looks like a pointer.

I am now forgetting what's that extra `%1` argument when it was initially creating that ccall? (Compare `stmt pre` and `stmt post`, which come from `update_statement!`.)