@KristofferC 2019-03-27
The way I understand it is:

```
julia> f() = ccall(:foo, Cvoid, (Ptr{Float64},Ptr{Int64}), A, B)
f (generic function with 1 method)

julia> @code_lowered f()
CodeInfo(
1 ─ %1 = (Core.apply_type)(Main.Ptr, Main.Float64)
│   %2 = (Base.cconvert)(%1, Main.A)
│   %3 = (Core.apply_type)(Main.Ptr, Main.Int64)
│   %4 = (Base.cconvert)(%3, Main.B)
│   %5 = (Core.apply_type)(Main.Ptr, Main.Float64)
│   %6 = (Base.unsafe_convert)(%5, %2)
│   %7 = (Core.apply_type)(Main.Ptr, Main.Int64)
│   %8 = (Base.unsafe_convert)(%7, %4)
│   %9 = $(Expr(:foreigncall, :(:foo), Nothing, svec(Ptr{Float64}, Ptr{Int64}), :(:ccall), 2, :(%6), :(%8), :(%4), :(%2)))
└──      return %9
)
```

The arguments `%4` and `%2` here are the result of `cconvert` and are only part of the `foreigncall` such that the GC will not think they are unused and free them.
What I think is the correct thing to do is to look at `%4, then go one step backwards and take the second argument from the `cconvert` ( `Main.B`) and that is the argument we should pass to the ccall-wrapper function. Right now, I am directly passing the `unsafe_convert´ value which is not right.