@KristofferC 2019-03-27
However, the GC things don't seem to always be there:

```
julia> f(x) =  ccall(:jl_value_ptr, Ptr{Cvoid}, (Any, Ptr{Int64}), x, y)
f (generic function with 2 methods)

julia> @code_lowered f(5)
CodeInfo(
1 ─ %1 = (Core.apply_type)(Main.Ptr, Main.Int64)
│   %2 = (Base.cconvert)(%1, Main.y)
│   %3 = (Core.apply_type)(Main.Ptr, Main.Int64)
│   %4 = (Base.unsafe_convert)(%3, %2)
│   %5 = $(Expr(:foreigncall, :(:jl_value_ptr), Ptr{Nothing}, svec(Any, Ptr{Int64}), :(:ccall), 2, :(x), :(%4), :(%2)))
└──      return %5
)
```

So here we only call `cconvert` on `y`. Is it because `Any` is special cased to not have a `cconvert` on them?