@KristofferC 2019-03-27
Implemented the idea above. Getting most of the ways now but errors when stepping through test. The error is

```
julia> frame, pc = debug_command(frame, :s)
(Frame for print_to_string(xs...) in Base at strings/io.jl:113
⋮
  27 121  │   %27 = (tuple)(siz)
  28 121  │   %28 = (%26)(%27)
  29 121  │   %29 = (Core.kwfunc)(Base.GenericIOBuffer{Array{UInt8,1}})
  30 121  │         s = (%29)(%28, Base.GenericIOBuffer{Array{UInt8,1}})
  31 122  │   %31 = xs
⋮
xs = (:(a + b),)
siz = 8
#temp# = nothing
x = :(a + b)
caller: string(xs...) in Base at strings/io.jl:156, 29)

julia> frame, pc = debug_command(frame, :s)
ret = getfield(Core, Symbol("#kw#Type"))()
ERROR: AssertionError: ret isa BreakpointRef
Stacktrace:
 [1] maybe_step_through_wrapper!(::Any, ::Frame) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\commands.jl:253
 [2] debug_command(::Any, ::Frame, ::Symbol, ::Bool) at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\commands.jl:455
 [3] debug_command at C:\Users\Kristoffer\Debugging\JuliaInterpreter\src\commands.jl:480 [inlined] (repeats 2 times)
 [4] top-level scope at none:0
```