@timholy 2019-03-27
I'm not really sure, and `foreigncall` is not documented [here](https://docs.julialang.org/en/latest/devdocs/ast/#Expr-types-1). When i was working on the existing ccall implementation I struggled with this a lot but don't really remember the conclusions I came to. I suspect `Any` gets passed to C as a `jl_value_t*`?