@KristofferC 2019-03-27
I've gotten this passing by excluding some tricky cases which is when the library has to be looked up... Running the interpreter linearly as is done right now doesn't work because there might be non evaluatable statements in between the foreign call and the definition of the library. We should probably run the Julia test suite on this.

Some performance numbers:

Before:

```
julia>  s = "foo"; @btime @interpret pointer_from_objref(s)
  1.089 ms (503 allocations: 30.94 KiB)

julia>  str = "abc 123\n"^100; @btime @interpret split(str, '\n')
  1.123 s (811505 allocations: 38.22 MiB)

julia> using Plots; @time @interpret plot(rand(5))
 18.342900 seconds (20.03 M allocations: 876.301 MiB, 1.09% gc time)
```


After:

```
julia> s = "foo"; @btime @interpret pointer_from_objref(s)
  4.400 μs (44 allocations: 2.38 KiB)

julia> str = "abc 123\n"^100; @btime @interpret split(str, '\n')
  188.822 ms (442822 allocations: 15.85 MiB)

julia> using Plots; @time @interpret plot(rand(5))
  8.551581 seconds (24.95 M allocations: 1.043 GiB, 2.78% gc time)
```