@KristofferC 2019-03-27
Here is a standalone example of the sparam problem:

First, a function that does a ccall. We extract the `Ptr{T}` return type:

```jl
julia> my_unsafe_convert(::Type{Ptr{T}}, a::Array{T}) where {T} = ccall(:jl_array_ptr, Ptr{T}, (Any,), a)my_unsafe_convert (generic function with 1 method)

julia> c = (@code_lowered my_unsafe_convert(Ptr{Int}, [1,2])).code
2-element Array{Any,1}:
 :($(Expr(:foreigncall, :(:jl_array_ptr), Ptr{T}, svec(Any), :(:ccall), 1, :(_3))))
 :(return %1)

julia> ptr_t = c[1].args[2]
Ptr{T}
```

Then we create a new function, and interpolate that variable:

```jl
julia> @eval my_unsafe_convert_2(::Type{Ptr{T}}, a::Array{T}) where {T} = ccall(:jl_array_ptr, $(ptr_t), (Any,), a):

julia> my_unsafe_convert_2(Ptr{Int}, [1,2])
ERROR: ccall return type Ptr should have an element type (not Ptr{<:T})
Stacktrace:
 [1] my_unsafe_convert_2(::Type{Ptr{Int64}}, ::Array{Int64,1}) at .\REPL[24]:1
 [2] top-level scope at none:0
```

As can be seen, the `Ptr{T}` we interpolated doesn't "bind" correctly with the `where {T}`. The `T` is probably just a label that is now not matching and I don't know how to update the `Ptr{T}` to work in the second method.